Фикс тикток добавление id в env

+++ Bot/handlers/dowload_movies.py
@@ -9,20 +9,29 @@ ydl_opts = {
     'format': 'mp4',
     'outtmpl': 'downloaded_video.%(ext)s',
 }
+ALLOWED_URLS = (
+    'https://www.instagram.com/reel/',
+    'https://www.tiktok.com/',
+    'https://vt.tiktok.com/',
+    'https://vm.tiktok.com/',
+)
 
 async def download_video(update: Update, context: ContextTypes.DEFAULT_TYPE):
     url = update.message.text
     chat_id = update.message.chat_id
     user = update.effective_user
     username = user.username if user.username else user.first_name
+    username_resent = update.message.api_kwargs.get("forward_from")
 
     if is_allowed_chat(chat_id):
-        if url.startswith((
-                'https://www.instagram.com/reel/',
-                'https://www.tiktok.com/',
-                'https://vt.tiktok.com/'
-        )):
+        if url.startswith(ALLOWED_URLS):
             print(f"Запрос на скачивание: {url}")
+            message_to_telegram = f"🎬 Отправлено пользователем: @{username}"
+
+            if username_resent:
+                message_to_telegram += f"\n📌 Переслано от пользователя: @{username_resent.get("username") or username_resent.get("first_name")}"
+
+            message_to_telegram += f"\n🌐 Ссылка: {url[:1000]}"
 
             await context.bot.send_chat_action(chat_id, "upload_video")
 
@@ -35,7 +44,7 @@ async def download_video(update: Update, context: ContextTypes.DEFAULT_TYPE):
                 with open(video_file, 'rb') as video:
                     await context.bot.send_video(
                         chat_id, video,
-                        caption=f"🎬 Отправлено пользователем: @{username}"
+                        caption=message_to_telegram
                     )
 
                 # Удаление файла после отправки
@@ -43,7 +52,10 @@ async def download_video(update: Update, context: ContextTypes.DEFAULT_TYPE):
                 await context.bot.delete_message(chat_id, update.message.message_id)
 
             except Exception as e:
-                print(f"Ошибка скачивания: {e}")
-                await context.bot.send_message(chat_id, f"⚠️ Поддерживаются только Instagram и TikTok. Отправь @{username} нормальную ссылку, а не это: {url[:1000]}")
-                # Удаляем неподдерживаемое сообщение
-                await context.bot.delete_message(chat_id, update.message.message_id)
+                if 'photo' in e:
+                    print(f"Ошибка скачивания: {e}")
+                else:
+                    print(f"Ошибка скачивания: {e}")
+                    await context.bot.send_message(chat_id, f"⚠️ Поддерживаются только Instagram и TikTok. Отправь @{username} нормальную ссылку, а не это: {url[:1000]} \n Ошибка: {e}")
+                    # Удаляем неподдерживаемое сообщение
+                    await context.bot.delete_message(chat_id, update.message.message_id)
